Fix runtime errors cause by stupidity
Confusing myself

+++ src/main/java/edu/uw/citw/persistence/repository/InstanceParticipantsRepository.java
@@ -14,5 +14,5 @@ import java.util.List;
 public interface InstanceParticipantsRepository extends CrudRepository<InstanceParticipant, Long> {
 
     @Query(value = "select * from laugh_instance_participants where instance_id = ?1", nativeQuery = true)
-    List<InstanceParticipant> findByInstanceId(Integer id);
+    List<InstanceParticipant> findByInstanceId(Long id);
 }
